requirements.txt
pyyaml~=6.0.2
python-json-logger~=3.2.1
